+++ settingrepo-sdk/src/main/java/com/dwarfeng/settingrepo/sdk/bean/dto/WebInputIahnNodeLocaleRemoveInfo.java
@@ -1,10 +1,11 @@
 package com.dwarfeng.settingrepo.sdk.bean.dto;
 
 import com.alibaba.fastjson.annotation.JSONField;
-import com.dwarfeng.settingrepo.sdk.util.Constraints;
+import com.dwarfeng.settingrepo.sdk.util.ValidIahnNodeCountry;
+import com.dwarfeng.settingrepo.sdk.util.ValidIahnNodeLanguage;
+import com.dwarfeng.settingrepo.sdk.util.ValidIahnNodeVariant;
 import com.dwarfeng.settingrepo.stack.bean.dto.IahnNodeLocaleRemoveInfo;
 import com.dwarfeng.subgrade.stack.bean.dto.Dto;
-import org.hibernate.validator.constraints.Length;
 
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
@@ -45,20 +46,15 @@ public class WebInputIahnNodeLocaleRemoveInfo implements Dto {
     private String[] args;
 
     @JSONField(name = "language")
-    @NotNull
-    @NotEmpty
-    @Length(min = Constraints.LENGTH_IAHN_NODE_LANGUAGE_MIN, max = Constraints.LENGTH_IAHN_NODE_LANGUAGE_MAX)
+    @ValidIahnNodeLanguage
     private String language;
 
     @JSONField(name = "country")
-    @NotNull
-
-    @Length(min = Constraints.LENGTH_IAHN_NODE_COUNTRY_MIN, max = Constraints.LENGTH_IAHN_NODE_COUNTRY_MAX)
+    @ValidIahnNodeCountry
     private String country;
 
     @JSONField(name = "variant")
-    @NotNull
-    @Length(min = Constraints.LENGTH_IAHN_NODE_VARIANT_MIN, max = Constraints.LENGTH_IAHN_NODE_VARIANT_MAX)
+    @ValidIahnNodeVariant
     private String variant;
 
     public WebInputIahnNodeLocaleRemoveInfo() {